part 1 of introducing Seconds data type

+++ op-challenger/game/monitor.go
@@ -129,7 +129,7 @@ func (m *gameMonitor) progressGames(ctx context.Context, blockHash common.Hash, 
 }
 
 func (m *gameMonitor) onNewL1Head(ctx context.Context, sig eth.L1BlockRef) {
-	m.clock.SetTime(sig.Time)
+	m.clock.SetTime(uint64(sig.Time))
 	if err := m.progressGames(ctx, sig.Hash, sig.Number); err != nil {
 		m.logger.Error("Failed to progress games", "err", err)
 	}
